iteration2/requirements.txt
pytest>=7.0.0
mypy>=1.0.0
pyyaml>=6.0